% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-mapping.R
\name{compound2cid}
\alias{compound2cid}
\title{compound2cid}
\usage{
compound2cid(medam, compound)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db}

\item{compound}{Synonyms of compound (ignored case), e.g. "aspirin" or
some external id (see details).}
}
\value{
a tibble contained query compound, PubChem cid and STITCH cid.
}
\description{
Convert synonyms of compound to PubChem cid and STITCH cid
(which are are derived from PubChem).
}
\details{
some external id of "aspirin":
\itemize{
\item \code{KEGG compound} D00109
\item \code{ChEBI} CHEBI:15365
\item \code{HMDB} HMDB0001879
\item \code{CAS-RN} 50-78-2
\item etc.
}
}
\examples{
\dontrun{
# remotes::install_github("rstudio/pool")
library(pool)

medamdb <- dbPool(drv = RSQLite::SQLite(), dbname = "MeDAM.db")
metabolites <- c("aspirin", "D00109", "CHEBI:15365", "HMDB0001879")
c2cid <- compound2cid(medamdb, metabolites)
poolClose(medamdb)
}
}
