% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-utilities.R
\name{dbquery}
\alias{dbquery}
\title{Query database method}
\usage{
dbquery(dbconn, dbtbl, ...)
}
\arguments{
\item{dbconn}{A Pool object connecting database}

\item{dbtbl}{A table in the database}

\item{...}{The arguments passed on to \code{\link[dplyr]{filter}} and allows
you to do queries using dplyr that can be transformed into a SELECT SQL
statement.}
}
\value{
tibble object
}
\description{
Query a database using dplyr and pool.
}
\examples{
\dontrun{
# remotes::install_github("rstudio/pool")
library(pool)
library(dplyr)

medamdb <- dbPool(drv = RSQLite::SQLite(), dbname = "MeDAM.db")
disease <- c("Pre-eclampsia", "Endometriosis")
medamdb |>
  dbquery("disease2doid", disease \%in\% !!disease)
poolClose(medamdb)
}
}
