% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doid-associated-cid.R
\name{drcid_search}
\alias{drcid_search}
\title{Search disease-related compounds}
\usage{
drcid_search(medam, cid, doid)
}
\arguments{
\item{medam}{a Pool object connected MeDAM.db}

\item{cid}{PubChem cid, e.g. '2244'.}

\item{doid}{Disease Ontology ID (DOID), e.g. 'DOID:178'.}
}
\value{
A table in which each row represents the compound-disease (cid-doid)
association.
}
\description{
Retrieve disease-related compounds.
}
\details{
Disease-related compounds collected from Human Metabolome Database
(HMDB) and Comparative Toxicogenomics Database (CTD).
}
\examples{
\dontrun{
# remotes::install_github("rstudio/pool")
library(pool)
library(dplyr)

medamdb <- dbPool(drv = RSQLite::SQLite(), dbname = "MeDAM.db")
cid <- c("1", "3715")
doid <- c("DOID:83", "DOID:3021")
cid2doid <- drcid_search(medamdb, cid, doid)
poolClose(medamdb)
}
}
