% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease-ora.R
\name{drgene_ora}
\alias{drgene_ora}
\title{Disease-related genes ORA}
\usage{
drgene_ora(nodes, drg, universe)
}
\arguments{
\item{nodes}{The tibble \code{nodes} from STRING or STITCH network.}

\item{drg}{Disease-related genes (entrez gene id).}

\item{universe}{Count of universal background protein-coding genes.
For example, background of homo sapiens is 21306.}
}
\value{
A list contained:
\itemize{
\item \code{ngene} number of interaction genes (proteins) in "nodes"
\item \code{indrg} number of intersection between genes (proteins) in "nodes" and
disease-related genes.
\item \code{pvalue} p value of hypergeometric test
\item \code{overlap} overlap genes between genes (proteins) in "nodes" and
disease-related genes
}
}
\description{
Over-representation analysis (ORA) for disease-related genes.
Ref: fig 1.C in Huimin Zheng, et al. (2022).
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
