% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease-ora.R
\name{drgene_search}
\alias{drgene_search}
\title{Search disease-related genes}
\usage{
drgene_search(medam, doid)
}
\arguments{
\item{medam}{a Pool object connected MeDAM.db}

\item{doid}{Disease Ontology ID (DOID).}
}
\value{
entrez gene id
}
\description{
Retrieve disease-related genes.
}
\details{
The aliases of disease can be converted to DOID using
\code{\link{disease2doid}}.
}
\examples{
\dontrun{
# remotes::install_github("rstudio/pool")
library(pool)
library(dplyr)

medamdb <- dbPool(drv = RSQLite::SQLite(), dbname = "MeDAM.db")
pe_doid <- disease2doid(medamdb, "pre-eclampsia")
pe_doid <- pull(pe_doid, doid) ## DOID:10591
pe_drg <- drgene_search(medamdb, pe_doid)
poolClose(medamdb)
}
}
