% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Enrichment analysis}
\usage{
enrichment(
  nodes,
  padj = "BH",
  pcutoff = 0.05,
  ont = c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO")
)
}
\arguments{
\item{nodes}{The tibble \code{nodes} from STRING or STITCH network.}

\item{padj}{P-values adjusted methods, one of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none".}

\item{pcutoff}{0.05.}

\item{ont}{c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO").}
}
\value{
A \code{\link[DOSE]{enrichResult-class}} instance for GO-BP, GO-MF,
GO-CC, KEGG, WikiPathways, Reactome and DO (Disease Ontology) enrichment
analysis.
}
\description{
Functional or pathway enrichment analysis for genes/proteins
interactied with metabolites. Ref: fig 1.C in Huimin Zheng, et al. (2022).
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
