% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeDAM-package.R
\docType{package}
\name{MeDAM-package}
\alias{MeDAM}
\alias{MeDAM-package}
\title{MeDAM: extrapolation of potential Metabolite-Disease Associations by Mining biomedical knowledge}
\description{
MeDAM is a complementary bioinformatic pipeline, which is base on biomedical knowledge discovery to effectively infer the potential functions of understudied metabolites associated with diseases for further investigation.
}
\author{
\strong{Maintainer}: Li Zhan \email{smu18575877413@gmail.com} (\href{https://orcid.org/0009-0003-7470-7586}{ORCID})

}
\keyword{internal}
