% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment-analysis.R
\name{medam_add_enrichment}
\alias{medam_add_enrichment}
\title{Add result of enrichment analysis}
\usage{
medam_add_enrichment(
  medamres,
  metabolite,
  branch,
  padj = "BH",
  pcutoff = 0.05,
  ont = c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO")
)
}
\arguments{
\item{medamres}{Result of MeDAM batch search.}

\item{metabolite}{Metabolite name.}

\item{branch}{MeDAM branch.}

\item{padj}{P-values adjusted methods, one of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none".}

\item{pcutoff}{0.05.}

\item{ont}{c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO").}
}
\value{
Result of MeDAM batch search.
}
\description{
Add result of enrichment analysis into the slot \verb{$enriment}
in result of MeDAM batch search when select a metabolite and its branch.
}
