% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medam-batch.R
\name{medam_batch_manual}
\alias{medam_batch_manual}
\title{MeDAM batch search for biomarker metabolites you input manually}
\usage{
medam_batch_manual(
  medam,
  metabolites,
  disease,
  abundance = NULL,
  score = 700,
  ecpi_score = 800,
  networkType = "signed",
  corMethod = "spearman",
  minModuleSize = 5,
  cutHeight = 0.25,
  universe = 21306,
  nthreads = 10
)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db.}

\item{metabolites}{A biomarker metabolites list.}

\item{disease}{Aliases of disease (ignored case),  e.g. "pre-eclampsia".}

\item{abundance}{Abundance table, a matrix or data frame in which columns are
metabolites and rows ar samples. It will execute the MeDAM batch search for
co-abundant metabolites when the abundance table is not NULL.}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (default: 700).}

\item{ecpi_score}{Threshold of significance for metabilite-protein
experimental interactions (from 0 to 1000, default: 800). See
\code{stitch_expt_cpi()}.}

\item{networkType}{Network type, one of "signed", "unsigned" and
"signed hybrid".}

\item{corMethod}{Correlation algorithm for network construction, e.g.,
"spearman" or "pearson".}

\item{minModuleSize}{Minimum size of module or cluster, default is 5.}

\item{cutHeight}{Maximum dissimilarity (i.e., 1-correlation) that qualifies
modules for merging.}

\item{universe}{Count of universal background protein-coding genes.
For example, background of homo sapiens is 21306.}

\item{nthreads}{Default 10.}
}
\value{
a list contained:
\itemize{
\item \code{daa} The metabolites you input and the module to which they belong
using \code{\link{wgcna_analysis}} while the abundance table is not NULL.
\item \code{network} Interaction network. See \code{\link{string_network}} and
\code{\link{stitch_network}}.
\item \code{drgora} Disease-related genes ORA. See \code{\link{drgene_ora}}
\item \code{disease} Common name, DOID and related (entrez) genes of disease.
}
}
\description{
Batch search for potential metabolite-disease associations using
MeDAM pipeline. Ref: fig 1.C in Huimin Zheng, et al. (2022).
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
