% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medam-arm.R
\name{medam_coab_metabolites}
\alias{medam_coab_metabolites}
\title{The co-abundance metabolites branch in MeDAM pipeline}
\usage{
medam_coab_metabolites(
  medam,
  compound,
  score = 700,
  use_wgcna = TRUE,
  abundance = NULL,
  networkType = "signed",
  corMethod = "spearman",
  minModuleSize = 5,
  cutHeight = 0.25,
  drg = NULL,
  universe = 21306,
  padj = "BH",
  pcutoff = 0.05,
  ont = c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO")
)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db}

\item{compound}{A single target compound or a cluster of compounds correlated
with abundance (i.e., consistency module).}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (default: 700).}

\item{use_wgcna}{\code{TRUE}, see details.}

\item{abundance}{Abundance table, a matrix or data frame in which columns are
metabolites and rows ar samples.}

\item{networkType}{Network type, one of "signed", "unsigned" and
"signed hybrid".}

\item{corMethod}{Correlation algorithm for network construction, e.g.,
"spearman" or "pearson".}

\item{minModuleSize}{Maximum size of module or cluster, default is 5.}

\item{cutHeight}{Maximum dissimilarity (i.e., 1-correlation) that qualifies
modules for merging.}

\item{drg}{Disease-related genes(entrez gene id).}

\item{universe}{Count of universal background protein-coding genes.
For example, background of homo sapiens is 21306.}

\item{padj}{P-values adjusted methods, one of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none".}

\item{pcutoff}{0.05.}

\item{ont}{c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO").}
}
\value{
list contained weights of each edge and annotations of each node
as following:

the tibble \code{edges} with fields:
\itemize{
\item \code{node1, node2} protein- or compound-nodes in the network
\item \code{type} type of interaction edges in network, e.g., cpi
(compound-protein), cci (compound-compound) and ppi (protein-protein)
\item \code{score} combined score
\item \code{experiment} experimental score
\item \code{neighborhood} gene neighborhood score
\item \code{fusion} gene fusion score
\item \code{prediction} prediction score
\item \code{cooccurence} cooccurence score
\item \code{coexpression} coexpression score
\item \code{database} database score
\item \code{textmining} textmining score
\item \code{evidence} more convincing interaction, the rank are: 1) experiment,
}
\enumerate{
\item neighborhood/fusion, 3) prediction/cooccurence/coexpression,
\item database/textmining
}

the tibble \code{nodes} with fields:
\itemize{
\item \code{id} STRING id (compound) or STTICH id (protein)
\item \code{type} type of nodes in network, protein or metabolite
\item \code{name} common synonyms of compound or protein
\item \code{external} external id of compound (PubChem cid) or protein
(entrez gene id)
\item \code{annotion} description of compound or protein
}
}
\description{
the branch 1 of second arm of pipeline in medam. Ref: fig 1.C in
Huimin Zheng, et al. (2022).
}
\details{
Use \code{wgcna_analysis()} to calculate consistency modules
(co-abundance compounds) and select the module contained target compound
while length of \code{compound} is 1, \code{use_wgcna} is TRUE and
\code{abundance} is not null.
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
\seealso{
\code{\link{wgcna_analysis}}, \code{\link{stitch_network}}
}
