% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medam-arm.R
\name{medam_ssim_metabolites}
\alias{medam_ssim_metabolites}
\title{The structural similar metabolites branch in MeDAM pipeline}
\usage{
medam_ssim_metabolites(
  medam,
  compound,
  score = 700,
  search_ssim = TRUE,
  drg = NULL,
  universe = 21306,
  padj = "BH",
  pcutoff = 0.05,
  ont = c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO")
)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db}

\item{compound}{A single target compound or a set of compounds shared similar
structure.}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (default: 700).}

\item{search_ssim}{\code{TRUE}, see details.}

\item{drg}{Disease-related genes(entrez gene id).}

\item{universe}{Count of universal background protein-coding genes.
For example, background of homo sapiens is 21306.}

\item{padj}{P-values adjusted methods, one of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none".}

\item{pcutoff}{0.05.}

\item{ont}{c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO").}
}
\value{
list contained weights of each edge, annotations of each node
result of disease-related genes ORA analysis and result of enrichment
analysis as following:

the tibble \code{edges} with fields:
\itemize{
\item \code{node1, node2} protein- or compound-nodes in the network
\item \code{type} type of interaction edges in network, e.g., cpi
(compound-protein), cci (compound-compound) and ppi (protein-protein)
\item \code{score} combined score
\item \code{experiment} experimental score
\item \code{neighborhood} gene neighborhood score
\item \code{fusion} gene fusion score
\item \code{prediction} prediction score
\item \code{cooccurence} cooccurence score
\item \code{coexpression} coexpression score
\item \code{database} database score
\item \code{textmining} textmining score
\item \code{evidence} more convincing interaction, the rank are: 1) experiment,
}
\enumerate{
\item neighborhood/fusion, 3) prediction/cooccurence/coexpression,
\item database/textmining
}

the tibble \code{nodes} with fields:
\itemize{
\item \code{id} STRING id (compound) or STTICH id (protein)
\item \code{type} type of nodes in network, protein or metabolite
\item \code{name} common synonyms of compound or protein
\item \code{external} external id of compound (PubChem cid) or protein
(entrez gene id)
\item \code{description} description of compound or protein
}
}
\description{
The branch 2 of second arm of pipeline in medam. Ref: fig 1.C in
Huimin Zheng, et al. (2022).
}
\details{
Use \code{ssim_search()} to search a set of compounds with similar
structure for target compound while length of \code{compound} is 1 and
\code{use.ssim_search} is TRUE.
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
\seealso{
\code{\link{ssim_search}}, \code{\link{stitch_network}}
}
