% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medam-arm.R
\name{medam_target_proteins}
\alias{medam_target_proteins}
\title{The target proteins branch in MeDAM pipeline}
\usage{
medam_target_proteins(
  medam,
  compound,
  protein = NULL,
  score = 700,
  use_ecpi = TRUE,
  ecpi_score = 800,
  drg = NULL,
  universe = 21306,
  padj = "BH",
  pcutoff = 0.05,
  ont = c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO")
)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db.}

\item{compound}{Target compound, e.g. Ononetin.}

\item{protein}{Target proteins regulated by compound and its intermediates,
e.g. TRPM3 (be inhibited by Ononetin) and TRPA1(be activated).}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (default: 700).}

\item{use_ecpi}{\code{TRUE}, see details.}

\item{ecpi_score}{Threshold of significance for metabilite-protein
experimental interactions (from 0 to 1000, default: 800).}

\item{drg}{Disease-related genes(entrez gene id).}

\item{universe}{Count of universal background protein-coding genes.
For example, background of homo sapiens is 21306.}

\item{padj}{P-values adjusted methods, one of "holm", "hochberg", "hommel",
"bonferroni", "BH", "BY", "fdr", "none".}

\item{pcutoff}{0.05.}

\item{ont}{c("BP", "MF", "CC", "KEGG", "WikiPathways", "Reactome", "DO").}
}
\value{
list contained weights of each edge and annotations of each node
as following:

the tibble \code{edges} with fields:
\itemize{
\item \code{node1, node2} protein- or compound-nodes in the network
\item \code{type} type of interaction edges in network, e.g., cpi
(compound-protein), cci (compound-compound) and ppi (protein-protein)
\item \code{score} combined score
\item \code{experiment} experimental score
\item \code{neighborhood} gene neighborhood score
\item \code{fusion} gene fusion score
\item \code{prediction} prediction score
\item \code{cooccurence} cooccurence score
\item \code{coexpression} coexpression score
\item \code{database} database score
\item \code{textmining} textmining score
\item \code{evidence} more convincing interaction, the rank are: 1) experiment,
}
\enumerate{
\item neighborhood/fusion, 3) prediction/cooccurence/coexpression,
\item database/textmining
}

the tibble \code{nodes} with fields:
\itemize{
\item \code{id} STRING id (compound) or STTICH id (protein)
\item \code{type} type of nodes in network, protein or metabolite
\item \code{name} common synonyms of compound or protein
\item \code{external} external id of compound (PubChem cid) or protein
(entrez gene id)
\item \code{annotion} description of compound or protein
}
}
\description{
The first arm of pipeline in medam. Ref: fig 1.C in
Huimin Zheng, et al. (2022).
}
\details{
In general, the \code{protein} obtained from literature and database
meant a increase in work. So, we recommend to search protein regulated by
target compound using \code{stitch_expt_cpi()} (STITCH metabilite-protein
experimental interactions network) required a \code{ecpi_score} (e.g. 700).
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
\seealso{
\code{\link{stitch_expt_cpi}}, \code{\link{string_network}}
}
