% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-utilities.R
\name{medamdbPool}
\alias{medamdbPool}
\title{MeDAM dbPool connections}
\usage{
medamdbPool(dbname)
}
\arguments{
\item{dbname}{MeDAM.db}
}
\value{
SQLiteConnection objects
}
\description{
Create a pool of database connections for MeDAM.db.
}
\examples{
\dontrun{
medamdb <- medamdbPool(dbname = "MeDAM.db")
}
}
