% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{PE.metabolomics}
\alias{PE.metabolomics}
\title{(Data) Non-targeted metabolomics for pre-eclampsia}
\description{
Contained two datasets, abundance and  metadata.
\itemize{
\item \code{abundance} contained 60 samples (nrow) and 8407 metabolites (ncol),
values have been log10-transformed.
\item \code{metadata} contained 30 patients with pre-eclampsia (PE) and
30 age-matched normotensive pregnant women (Ctrl) in their third trimester.
Ref: fig 1.C in Huimin Zheng, et al. (2022).
}
}
\examples{
data(PE.metabolomics)
abundance <- PE.metabolomics$abundance
metadata <- PE.metabolomics$metadata
unique(metadata$Statue2)
## [1] "normotension" "preeclampsia"
}
\references{
Huimin Zheng, et al. (2022). In silico method to maximise the
biological potential of understudied metabolomic biomarkers: a study in
pre-eclampsia.
}
\keyword{data}
