% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id-mapping.R
\name{protein2stringid}
\alias{protein2stringid}
\title{protein2stringid}
\usage{
protein2stringid(medam, protein)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db.}

\item{protein}{Synonyms of protein(ignored case), e.g. "TP53".}
}
\value{
A tibble contained query protein and STRING id.
}
\description{
Convert synonyms of protein to STRING id.
}
\examples{
\dontrun{
# remotes::install_github("rstudio/pool")
library(pool)

medamdb <- dbPool(drv = RSQLite::SQLite(), dbname = "MeDAM.db")
proteins <- c("PTCH1", "TP53", "BRCA1", "BRCA2")
protein2stringid(medamdb, proteins)
poolClose(medamdb)
}
}
