% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-methods.R
\name{scale_abundance}
\alias{scale_abundance}
\title{Scaling abundance table}
\usage{
scale_abundance(xMN, scaleC)
}
\arguments{
\item{xMN}{Abundance table, a matrix or data frame in which columns are
metabolites and rows ar samples.}

\item{scaleC}{Scaling methods for "abundance", one of "none", "center",
"pareto", and "standard" (default).}
}
\description{
Wrapper for "X-Scaling" in R pkg "ropls".
}
