% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{stitch_expt_cpi}
\alias{stitch_expt_cpi}
\title{STITCH experimental CPI}
\usage{
stitch_expt_cpi(medam, stitchid, score = 900)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db.}

\item{stitchid}{STITCH id, e.g. CIDs00002244.}

\item{score}{Threshold of experimental significance(default: 900).}
}
\value{
the tibble with fields:
\itemize{
\item \code{stitchv5} query metabilite(STITCH id in version V5)
\item \code{experiment} experimental-interaction score
\item \code{stringv10p5} experimental-interaction protein(STRING id in V10.5)
\item \code{stringv12} covert STRING id from V10.5 to V12
\item \code{description} description for protein
}
}
\description{
STITCH experimental metabilite-protein interactions.
}
\details{
The synonyms of compound can be converted to PubChem cid using
\code{\link{compound2cid}}.
}
