% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{string_network}
\alias{string_network}
\title{STRING network}
\usage{
string_network(medam, stringid, score = 700)
}
\arguments{
\item{medam}{A Pool object connected MeDAM.db.}

\item{stringid}{STRING id, e.g. ENSP00000269305.}

\item{score}{Threshold of significance to include a interaction, a number
between 0 and 1000 (default: 700).}
}
\value{
list contained weights of each edge and the annotations of each node
as following:

the tibble \code{edges} with fields:
\itemize{
\item \code{node1, node2} protein- or compound-nodes in the network
\item \code{type} type of interaction edges in network, e.g., cpi
(compound-protein), cci (compound-compound) and ppi (protein-protein)
\item \code{score} combined score
\item \code{experiment} experimental score
\item \code{neighborhood} gene neighborhood score
\item \code{fusion} gene fusion score
\item \code{prediction} prediction score
\item \code{cooccurence} cooccurence score
\item \code{coexpression} coexpression score
\item \code{database} database score
\item \code{textmining} textmining score
\item \code{evidence} more convincing interaction, the rank are: 1) experiment,
}
\enumerate{
\item neighborhood/fusion, 3) prediction/cooccurence/coexpression,
\item database/textmining
}

the tibble \code{nodes} with fields:
\itemize{
\item \code{id} STRING id (compound) or STTICH id (protein)
\item \code{type} type of nodes in network, protein or metabolite
\item \code{name} common synonyms of compound or protein
\item \code{external} external id of compound (PubChem cid) or protein
(entrez gene id)
\item \code{annotion} description of compound or protein
}
}
\description{
STRING interactions network.
}
\details{
The synonyms of protein can be converted to STRING id using
\code{\link{protein2stringid}}.
}
