% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{wgcna_analysis}
\alias{wgcna_analysis}
\title{WGCNA modules}
\usage{
wgcna_analysis(
  abundance,
  networkType = "signed",
  corMethod = "spearman",
  minModuleSize = 5,
  cutHeight = 0.25,
  nthreads = 10
)
}
\arguments{
\item{abundance}{Abundance table, a matrix or data frame in which columns are
metabolites and rows ar samples.}

\item{networkType}{Network type, one of "signed", "unsigned" and
"signed hybrid".}

\item{corMethod}{Correlation algorithm for network construction, e.g.,
"spearman" or "pearson"}

\item{minModuleSize}{Minimum size of module or cluster, default is 5.}

\item{cutHeight}{Maximum dissimilarity (i.e., 1-correlation) that qualifies
modules for merging.}

\item{nthreads}{Default 10.}
}
\value{
A \code{tibble} combined metabolites and modules.
}
\description{
Identify abundance-correlated metabolites (i.e., consistency
abundance module) using weighted correlation network analysis (WGCNA).
}
\examples{
\dontrun{
data(PE.metabolomics)
abun <- PE.metabolomics$abundance
wgcna_mods <- wgcna_analysis(abun)
}
}
