% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doid-associated-cid.R
\name{drcid_search}
\alias{drcid_search}
\title{Search disease-related compounds}
\usage{
drcid_search(medam, cid, doid)
}
\arguments{
\item{medam}{a Pool object connected MeDAM.db}

\item{cid}{PubChem cid, e.g. '2244'.}

\item{doid}{Disease Ontology ID (DOID), e.g. 'DOID:178'.}
}
\description{
Retrieve disease-related compounds.
}
\details{
Disease-related compounds collected from Human Metabolome Database
(HMDB) and Comparative Toxicogenomics Database (CTD).
}
